feat(orchestrator-service): add eventbridge strategy of bus-connector
GH-3
