+++ services/orchestrator-service/src/services/index.ts
@@ -1,7 +1,6 @@
-export * from './tenant-provisioning-handler.service';
+export * from './tenant-provisioning-consumer.service';
 export * from './tier-details.service';
 export * from './aws-code-build.service';
-export * from './orchestrator.service';
-export * from './tenant-provisioning-success.handler';
-export * from './tenant-deployment.handler';
+export * from './tenant-provisioning-success-consumer.service';
+export * from './tenant-deployment-consumer.service';
 export * from './data-store.service';
