+++ services/orchestrator-service/src/services/tenant-deployment-consumer.service.ts
@@ -1,26 +1,27 @@
-import {TenantDeploymentHandler} from '@sourceloop/ctrl-plane-orchestrator-service';
-import {injectable, BindingScope, Provider, service} from '@loopback/core';
+import {DefaultEventTypes} from '@sourceloop/ctrl-plane-orchestrator-service';
+import {injectable, BindingScope, service} from '@loopback/core';
 import {AnyObject} from '@loopback/repository';
 import {IncomingMessage} from 'http';
 import * as https from 'https';
 import * as http from 'http';
 import * as crypto from 'crypto';
 import {DataStoreService} from './data-store.service';
+import {consumer, IConsumer, QueueType} from 'loopback4-message-bus-connector';
 
 @injectable({scope: BindingScope.TRANSIENT})
-export class TenantDeploymentProvider
-  implements Provider<TenantDeploymentHandler>
+@consumer
+export class TenantDeploymentConsumerService
+  implements IConsumer<AnyObject, string>
 {
+  event: DefaultEventTypes.TENANT_DEPLOYMENT =
+    DefaultEventTypes.TENANT_DEPLOYMENT;
+  queue: QueueType = QueueType.EventBridge;
   constructor(
     @service(DataStoreService)
     private readonly dataStoreService: DataStoreService,
   ) {}
 
-  value() {
-    return async (body: AnyObject) => this.handler(body);
-  }
-
-  private async handler(detail: AnyObject): Promise<void> {
+  async handle(detail: AnyObject): Promise<void> {
     console.log('Tenant Deployment: ', detail);
 
     const httpModule = this.getHttpModule(detail.API_ENDPOINT);
