+++ services/tenant-mgmt-service/package.json
@@ -75,7 +75,7 @@
     "@opentelemetry/sdk-trace-base": "^1.15.0",
     "@opentelemetry/sdk-trace-node": "^1.15.0",
     "@sourceloop/core": "^20.0.0",
-    "@sourceloop/ctrl-plane-tenant-management-service": "^1.0.1",
+    "@sourceloop/ctrl-plane-tenant-management-service": "^1.1.0",
     "@vendia/serverless-express": "^4.12.6",
     "auth": "^1.1.1",
     "auth0": "^4.10.0",
@@ -84,6 +84,7 @@
     "loopback-connector-postgresql": "^7.0.4",
     "loopback4-authentication": "^13.0.0",
     "loopback4-authorization": "^8.0.0",
+    "loopback4-message-bus-connector": "^2.0.0",
     "swagger-stats": "^0.99.5",
     "symlink-resolver": "0.2.1",
     "tslib": "^2.0.0"
