+++ services/tenant-mgmt-service/src/services/event.service.ts
@@ -4,9 +4,9 @@ import {
   SubscriptionDTO,
   TenantWithRelations,
 } from '@sourceloop/ctrl-plane-tenant-management-service';
-import {EventBridgeClient, PutEventsCommand} from '@aws-sdk/client-eventbridge';
 import {AnyObject} from '@loopback/repository';
 import {IEventConnector} from '@sourceloop/ctrl-plane-tenant-management-service';
+import {Producer, producer, QueueType} from 'loopback4-message-bus-connector';
 
 export enum Builder {
   CODE_BUILD = 'CODE_BUILD',
@@ -23,19 +23,14 @@ type EventBodyType = {
 
 @injectable({scope: BindingScope.TRANSIENT})
 export class EventConnector implements IEventConnector<EventBodyType> {
-  constructor() {}
+  constructor(
+    @producer(QueueType.EventBridge)
+    private eventBridgeProducer: Producer,
+  ) {}
 
   async publish(eventBody: EventBodyType) {
     console.log('Event body received in the event connector:', eventBody);
     const {type, secret, context, ...data} = eventBody;
-
-    console.log('Secret', secret);
-    console.log('Context', context);
-
-    // Configure the AWS SDK with your credentials and region
-    const eventBridgeClient = new EventBridgeClient({
-      region: process.env.EVENT_BUS_REGION,
-    });
 
     const extraPlanConfig: AnyObject = {};
 
@@ -54,48 +49,27 @@ export class EventConnector implements IEventConnector<EventBodyType> {
       'modifiedBy',
     ];
     data.tenant = this.removeFields(data.tenant, fieldsToRemove);
-
     // Define the event payload
     const eventPayload = {
-      Entries: [
-        {
-          Source: 'TenantManagementService',
-          DetailType: type,
-          Detail: JSON.stringify({
-            tenant: data.tenant,
-            appConfig: data?.appConfig,
-            planConfig: data.subscription.plan,
-            builderConfig: {
-              type: Builder.CODE_BUILD,
-              config: {
-                environmentOverride: {
-                  ...extraPlanConfig,
-                  tenant: JSON.stringify(data.tenant),
-                  secret: secret,
-                  context: context,
-                },
-              },
+      type,
+      data: {
+        tenant: data.tenant,
+        appConfig: data.appConfig,
+        planConfig: data.subscription.plan,
+        builderConfig: {
+          type: Builder.CODE_BUILD,
+          config: {
+            environmentOverride: {
+              ...extraPlanConfig,
+              tenant: JSON.stringify(data.tenant),
+              secret,
+              context,
             },
-          }),
-          EventBusName: process.env.EVENT_BUS_NAME,
+          },
         },
-      ],
+      },
     };
-
-    console.log('Event bus name', process.env.EVENT_BUS_NAME);
-    // Create the PutEventsCommand with the event payload
-    const putEventsCommand = new PutEventsCommand(eventPayload);
-
-    // Send the event to the event bus
-    const response = await eventBridgeClient
-      .send(putEventsCommand)
-      .catch(err => {
-        console.error('Error sending event:', err);
-      });
-
-    if (response) {
-      console.log('Event sent successfully:', response);
-    }
+    await this.eventBridgeProducer.send(eventPayload);
   }
 
   private removeFields(obj: AnyObject, fieldsToRemove: string[]): AnyObject {
